refactor error utils
